Added webpack and setup page to use webpack-dev-server as well as splitting the components into separate files to be included as modules.
