+++ package.json
@@ -5,12 +5,21 @@
   "main": "server.js",
   "scripts": {
     "stop": "echo not implemented",
-    "start": "node ./server.js"
+    "start": "node ./server.js",
+    "dev": "webpack-dev-server --progress --colors --port 5555"
   },
   "dependencies": {
     "express": "4.13.0",
+    "react": "^0.13.3",
     "redis": "0.12.1",
     "socket.io": "1.3.5",
     "socket.io-redis": "0.1.4"
+  },
+  "devDependencies": {
+    "babel-core": "^5.6.5",
+    "babel-loader": "^5.1.4",
+    "node-libs-browser": "^0.5.2",
+    "webpack": "^1.9.11",
+    "webpack-dev-server": "^1.9.0"
   }
 }
