+++ www/index.html
@@ -3,12 +3,11 @@
 <head>
   <title>Testing Grounds</title>
   <link rel="stylesheet" href="css/styles.css" type="text/css"/>
-  <script src="build/react.js"></script>
-  <script src="build/JSXTransformer.js"></script>
   <script src="/socket.io/socket.io.js"></script>
-  <script type="text/jsx" src="js/script.js"></script>
+  <script src="http://localhost:5555/webpack-dev-server.js"></script>
 </head>
 <body>
   <div id="Main"></div>
+  <script src="http://localhost:5555/assets/bundle.js"></script>
 </body>
 </html>
